Create tests for backend (#84)
* Create tests for backend

* Fix tests

* Fix linter issues

+++ backend/internal/stats/model.go
@@ -4,10 +4,10 @@ import "time"
 
 type MostUsedDomainResponse struct {
 	Domain string
-	Count  int
+	Count  int64
 }
 
 type ServerUsageByTimeRangeResponse struct {
 	TimeRange time.Time
-	Count     int
+	Count     int64
 }
